+++ canvas_sdk/effects/data_integration/__init__.py
@@ -1,15 +1,23 @@
 from .assign_document_reviewer import AssignDocumentReviewer as AssignDocumentReviewer
 from .assign_document_reviewer import Priority as Priority
 from .assign_document_reviewer import ReviewMode as ReviewMode
+from .base import _BaseDocumentEffect as _BaseDocumentEffect
+from .categorize_document import CategorizeDocument as CategorizeDocument
 from .junk_document import JunkDocument as JunkDocument
 from .link_document_to_patient import LinkDocumentToPatient as LinkDocumentToPatient
 from .prefill_document_fields import PrefillDocumentFieldData as PrefillDocumentFieldData
 from .prefill_document_fields import PrefillDocumentFields as PrefillDocumentFields
 from .prefill_document_fields import TemplateFields as TemplateFields
 from .remove_document_from_patient import RemoveDocumentFromPatient as RemoveDocumentFromPatient
+from .types import AnnotationItem as AnnotationItem
+from .types import DocumentType as DocumentType
 
 __all__ = __exports__ = (
+    "_BaseDocumentEffect",
+    "AnnotationItem",
     "AssignDocumentReviewer",
+    "CategorizeDocument",
+    "DocumentType",
     "PrefillDocumentFieldData",
     "JunkDocument",
     "LinkDocumentToPatient",
