+++ canvas_sdk/effects/data_integration/link_document_to_patient.py
@@ -2,11 +2,11 @@ from typing import Any
 
 from pydantic_core import InitErrorDetails
 
-from canvas_sdk.effects.base import EffectType, _BaseEffect
-from canvas_sdk.effects.data_integration.types import AnnotationItem
+from canvas_sdk.effects.base import EffectType
+from canvas_sdk.effects.data_integration.base import _BaseDocumentEffect
 
 
-class LinkDocumentToPatient(_BaseEffect):
+class LinkDocumentToPatient(_BaseDocumentEffect):
     """
     An Effect that links a document in the Data Integration queue to a patient
     using the patient's key.
@@ -35,48 +35,25 @@ class LinkDocumentToPatient(_BaseEffect):
             "patient_key",
         )
 
-    document_id: str | int | None = None
     patient_key: str | None = None
-    annotations: list[AnnotationItem] | None = None
-    source_protocol: str | None = None
 
     @property
     def values(self) -> dict[str, Any]:
         """The effect's values to be sent in the payload."""
         result: dict[str, Any] = {
-            "document_id": str(self.document_id).strip() if self.document_id is not None else None,
+            "document_id": self._serialize_document_id(),
             "patient_key": str(self.patient_key).strip() if self.patient_key else None,
         }
         if self.annotations is not None:
             result["annotations"] = self.annotations
         if self.source_protocol is not None:
-            result["source_protocol"] = self.source_protocol
+            result["source_protocol"] = self._serialize_source_protocol()
         return result
 
     def _get_error_details(self, method: Any) -> list[InitErrorDetails]:
         """Validate the effect fields and return any error details."""
         errors = super()._get_error_details(method)
-
-        # Validate document_id is non-empty if provided as string
-        if isinstance(self.document_id, str) and not self.document_id.strip():
-            errors.append(
-                self._create_error_detail(
-                    "value_error",
-                    "document_id must be a non-empty string",
-                    self.document_id,
-                )
-            )
-
-        # Validate patient_key is non-empty if provided
-        if self.patient_key is not None and not self.patient_key.strip():
-            errors.append(
-                self._create_error_detail(
-                    "value_error",
-                    "patient_key must be a non-empty string",
-                    self.patient_key,
-                )
-            )
-
+        errors.extend(self._validate_non_empty_string("patient_key", self.patient_key))
         return errors
