+++ canvas_sdk/effects/data_integration/remove_document_from_patient.py
@@ -2,10 +2,11 @@ from typing import Any
 
 from pydantic_core import InitErrorDetails
 
-from canvas_sdk.effects.base import EffectType, _BaseEffect
+from canvas_sdk.effects.base import EffectType
+from canvas_sdk.effects.data_integration.base import _BaseDocumentEffect
 
 
-class RemoveDocumentFromPatient(_BaseEffect):
+class RemoveDocumentFromPatient(_BaseDocumentEffect):
     """
     An Effect that removes/unlinks a document from a patient in the Data Integration queue.
 
@@ -25,14 +26,13 @@ class RemoveDocumentFromPatient(_BaseEffect):
         effect_type = EffectType.REMOVE_DOCUMENT_FROM_PATIENT
         apply_required_fields = ("document_id",)
 
-    document_id: str | int | None = None
     patient_id: str | None = None
 
     @property
     def values(self) -> dict[str, Any]:
         """The effect's values to be sent in the payload."""
         result: dict[str, Any] = {
-            "document_id": str(self.document_id).strip() if self.document_id is not None else None,
+            "document_id": self._serialize_document_id(),
         }
         if self.patient_id is not None:
             result["patient_id"] = self.patient_id.strip() if self.patient_id else None
@@ -41,27 +41,7 @@ class RemoveDocumentFromPatient(_BaseEffect):
     def _get_error_details(self, method: Any) -> list[InitErrorDetails]:
         """Validate the effect fields and return any error details."""
         errors = super()._get_error_details(method)
-
-        # Validate document_id is non-empty if provided as string
-        if isinstance(self.document_id, str) and not self.document_id.strip():
-            errors.append(
-                self._create_error_detail(
-                    "value_error",
-                    "document_id must be a non-empty string",
-                    self.document_id,
-                )
-            )
-
-        # Validate patient_id is non-empty if provided
-        if self.patient_id is not None and not self.patient_id.strip():
-            errors.append(
-                self._create_error_detail(
-                    "value_error",
-                    "patient_id must be a non-empty string",
-                    self.patient_id,
-                )
-            )
-
+        errors.extend(self._validate_non_empty_string("patient_id", self.patient_id))
         return errors
